ref(profiling): reorganize contexts
